t3_fastApi/pyproject.toml
[tool.poetry]
name = "t3-fastapi"
version = "0.1.0"
description = ""
authors = ["Данил Воронин <melinbodb@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.21"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
flask-admin = "^1.6.1"
flask-sqlalchemy = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
